BuyerApp -stripe entegration added

+++ mac_store_app/lib/views/screens/inner_screens/order_detail_screen.dart
@@ -44,7 +44,7 @@ class _OrderDetailScreenState extends State<OrderDetailScreen> {
 
     final productDoc = await productRef.get();
     if (productDoc.exists) {
-      int currentRating = productDoc.data()?['rating'] ?? 0;
+      double currentRating = productDoc.data()?['rating'] ?? 0;
 
       await productRef.update({
         'rating': currentRating + newRating,
